table impls restructuring
